#41 Move service packages, rename to DateRangeRequestParameters
