+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/service/employee/EmployeeService.java
@@ -1,4 +1,4 @@
-package com.senla.worklog.reminder.service;
+package com.senla.worklog.reminder.service.employee;
 
 import com.senla.worklog.reminder.dto.EmployeeDto;
