+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/service/worklogdebt/WorklogDebtsService.java
@@ -1,4 +1,4 @@
-package com.senla.worklog.reminder.service;
+package com.senla.worklog.reminder.service.worklogdebt;
 
 import com.senla.worklog.reminder.model.WorklogDebts;
 
@@ -9,9 +9,13 @@ import static java.time.DayOfWeek.MONDAY;
 import static java.time.LocalDate.now;
 
 public interface WorklogDebtsService {
+    WorklogDebts getAllForPeriod(LocalDate dateFrom, LocalDate dateTo);
+
+    default WorklogDebts getAllFrom(LocalDate dateFrom) {
+        return getAllForPeriod(dateFrom, now());
+    }
+
     default WorklogDebts getAllForCurrentWeek() {
         return getAllForPeriod(now().with(MONDAY), now().with(FRIDAY));
     }
-
-    WorklogDebts getAllForPeriod(LocalDate dateFrom, LocalDate dateTo);
 }
