+++ services/worklog-debts/src/main/java/com/senla/worklog/reminder/validator/DateRangeRequestParameters.java
@@ -18,7 +18,7 @@ import static org.springframework.format.annotation.DateTimeFormat.ISO.DATE;
 @NoArgsConstructor
 @AllArgsConstructor
 @DateFromMoreRecentThanDateTo(groups = Extended.class)
-public class GetAllDebtsForPeriodRequestParameters {
+public class DateRangeRequestParameters {
     @NotNull
     @PastOrPresent(message = "{dateFrom.pastOrPresent}")
     @DateTimeFormat(iso = DATE)
